+++ apps/backend/src/modules/vendors/libs/types/edx-course-response-dto.type.ts
@@ -1,0 +1,10 @@
+import { type EdxMedia } from "./edx-media.type.js";
+
+type EdxCourseResponseDto = {
+	course_id: string;
+	media: EdxMedia;
+	name: string;
+	short_description: string;
+};
+
+export { type EdxCourseResponseDto };