+++ apps/frontend/src/libs/types/async-thunk-config.type.ts
@@ -1,7 +1,9 @@
 import { type store } from "~/libs/modules/store/store.js";
 
+import { type AppDispatch } from "./app-dispatch.type.js";
+
 type AsyncThunkConfig = {
-	dispatch: typeof store.instance.dispatch;
+	dispatch: AppDispatch;
 	extra: typeof store.extraArguments;
 	state: ReturnType<typeof store.instance.getState>;
 };